% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPrices.R
\name{addPrices}
\alias{addPrices}
\title{Appending the USD prices to the dataset of transactions}
\usage{
addPrices(df, prices = df.prices)
}
\arguments{
\item{df}{a dataframe of transactions, result of getBTC()}

\item{prices}{a 2-column dataframe of daily BTC closing prices}
}
\value{
Returns a data frame
}
\description{
\code{addPrices} add an extra column to the transactions df with the USD BTC value of the transactions, calculated from daily closing prices
}
