% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Pairs the original address with addresses from the same wallet}
\usage{
cluster(df = data.frame(), address = c())
}
\arguments{
\item{df}{a dataframe obtained with getBTC}

\item{addresses}{a vector of addresses hash}
}
\value{
Returns a vector of addresses hash
}
\description{
\code{cluster} pairs the original vector of address(es) with siblings input addresses
}
