% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.address.transactions.details.R
\name{get.address.transactions.details}
\alias{get.address.transactions.details}
\title{Extracts the txs details from an address object}
\usage{
get.address.transactions.details(addr.list = list())
}
\arguments{
\item{addr.list}{an address list}
}
\value{
Returns a data.frame
}
\description{
\code{get.address.transactions.details} returns inputs, outputs, value, time, block height and ip relay for each transaction
}
