% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get}
\alias{get}
\title{Downloading and flattening ina df all transactions info from a set of BTC addresses}
\usage{
get(addresses = c())
}
\arguments{
\item{addresses}{a vector of addresses hash}
}
\value{
Returns a data frame
}
\description{
\code{get} is a wrapper - returns all transactions info of a vector of addresses from blockchain.info API.
}
