% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.scam.R
\name{get.scam}
\alias{get.scam}
\title{Checks whether the addresses have been reported on bitcoinwhoswho.com}
\usage{
get.scam(adds, apikey = "5d3a75b8-22ccf291-025f4e8d-739bfeaf")
}
\arguments{
\item{adds}{a vector of a single address hash}

\item{apikey}{bitcoiwhoswho.com API key}
}
\value{
Returns a dataframe or NULL
}
\description{
\code{get.scam} gets scam report date and description for the addresses, or returns NULL not report has been made.
}
