% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.transaction.details.R
\name{get.transaction.details}
\alias{get.transaction.details}
\title{Extracts the details of a single transaction}
\usage{
get.transaction.details(tx = list())
}
\arguments{
\item{tx}{an transaction list}
}
\value{
Returns a data.frame
}
\description{
\code{get.transaction.details} returns inputs, outputs, value, time, block height and ip relay for a single transaction
}
