% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klondikBTC-package.r
\docType{package}
\name{klondikBTC}
\alias{klondikBTC}
\alias{klondikBTC-package}
\title{klondikBTC.}
\description{
This package downloads and extracts transaction data from one or multiple BTC addresses from blockchain.info API.
}
