% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTransacs.R
\name{summaryTransacs}
\alias{summaryTransacs}
\title{Summarize a dataframe of transactions}
\usage{
summaryTransacs(df, addresses)
}
\arguments{
\item{df}{a dataframe obtained with getBTC}

\item{addresses}{a vector of addresses hash}
}
\value{
Returns a list
}
\description{
\code{summaryTransacs} returns number of transacitons, date of first and last transactions, transactions over $1000 and biggest transactions in which addresses are involved.
}
