% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.address.R
\name{get.address}
\alias{get.address}
\title{Downloads an address object from a single BTC address including offset transactions (when txs >50)}
\usage{
get.address(address = c(), offset = c())
}
\arguments{
\item{offset}{an integer that defines how many sets of 50 txs must be skipped backward starting from the last tx to date}

\item{addresses}{a vector of a single address hash}
}
\value{
Returns a list
}
\description{
\code{get.address} downloads address object from blockchain.info API.
}
