% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.address.R
\name{get.address}
\alias{get.address}
\title{Downloads an address object from a single BTC address including offset transactions (when txs >50)}
\usage{
get.address(address = c(), apikey)
}
\arguments{
\item{address}{a vector of a single address hash}

\item{apikey}{Blockchain.info API key}
}
\value{
Returns a list
}
\description{
\code{get.address} downloads address object from blockchain.info API.
}
